requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-docx
pdfminer.six
spacy
streamlit
python-dotenv

